stateful pause
Sufficient progress to push and reflect.

Still need to update Gen3 iteration, and eventually invert the logic flow so that the per-game iteration doesn't return an ienumerable, and instead iterates to the first satisfactory encounter (per the move/etc validation inside EncounterVerifier). Probably will be some code duplication, but it will be the minimal allocation approach for legality checking (checking go brrrr).

Much less IEnumerable usage in the encounter search, and no longer needs to have concatenated arrays of encounters since we can more finely control the progress of the state machine that is searching.

Inlines a lot of logic for GetPossible; it probably does cache a few more IEnumerables (branching), but these can eventually be optimized out like has been done for the finding match logic.

Removes explicit handling for Pressure slots in X/Y; just duplicate the flabebe form encounters with max level. Only adds a few more bytes to that pickle, in exchange for simpler code.

+++ PKHeX.Core/Legality/RNG/Frame/FrameGenerator.cs
@@ -18,9 +18,9 @@ public sealed class FrameGenerator
     public readonly bool Safari3;
 
     public Frame GetFrame(uint seed, LeadRequired lead) => new(seed, FrameType, lead);
-    public Frame GetFrame(uint seed, LeadRequired lead, uint esv, uint origin) => GetFrame(seed, lead, esv, esv, origin);
+    public Frame GetFrame(uint seed, LeadRequired lead, ushort esv, uint origin) => GetFrame(seed, lead, esv, esv, origin);
 
-    public Frame GetFrame(uint seed, LeadRequired lead, uint esv, uint lvl, uint origin) => new(seed, FrameType, lead)
+    public Frame GetFrame(uint seed, LeadRequired lead, ushort esv, ushort lvl, uint origin) => new(seed, FrameType, lead)
     {
         RandESV = esv,
         RandLevel = lvl,
